Finally sorted out the System.Data.dll reference framework issue as well as the npgsql.dll. They both needed to be copied into the assets folder for things to work correctly. Unity kept stripping out the reference from VS. They seem to be working under the Support Libraries folder. This may cause issues later as Unity is said to want them under the root of the Assets folder but that could be for older version of Unity.
Also moved  any closed source software into the "Closed Source Assets - Do Not Share!!!" folder and will and the contents of that folder to my .gitignore file. Still need to purge it from what ever is on GitHub before I open source this but it's a step in the right direction at least.